% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tract_conversions.R
\name{convert_2010_to_2020_bounds}
\alias{convert_2010_to_2020_bounds}
\title{redistributes 2010 data based on 2020 census boundaries
param: data(dataframe) -> data to redistribute. contains geoid(2010) and value attributes.
param: geoid_col(character - default="geoid") -> name of column with 2010 geoids
return: redistributed(dataframe) -> data redistributed to 2020 tract boundaries.
contains geoids(2020) and redistributed values}
\usage{
convert_2010_to_2020_bounds(data, geoid_col = "geoid", val_col = "value")
}
\description{
redistributes 2010 data based on 2020 census boundaries
param: data(dataframe) -> data to redistribute. contains geoid(2010) and value attributes.
param: geoid_col(character - default="geoid") -> name of column with 2010 geoids
return: redistributed(dataframe) -> data redistributed to 2020 tract boundaries.
contains geoids(2020) and redistributed values
}
