% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tract_conversions.R
\name{create_crosswalk}
\alias{create_crosswalk}
\title{generates a crosswalk file inclusive of all geography levels possible that
exist in geoids.
param: geoids(character vector) -> geoids to get crosswalk for.
return: dataframe of relationship data. columns:
geoid10, geoid20, area10, area20, area_part, type_change}
\usage{
create_crosswalk(geoids)
}
\description{
generates a crosswalk file inclusive of all geography levels possible that
exist in geoids.
param: geoids(character vector) -> geoids to get crosswalk for.
return: dataframe of relationship data. columns:
geoid10, geoid20, area10, area20, area_part, type_change
}
