% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tract_conversions.R
\name{get_2010_2020_bound_changes}
\alias{get_2010_2020_bound_changes}
\title{determines if/how census tract boundaries changed from 2010 to 2020. Adds
an additional "type_change" column to 2010-2020 tract relationship file
indicating whether a tract did not change ("same"), a tract was split but
bounds did not change ("split"), or a tracts bounds moved ("moved").
param: res(character - default "tract") -> resolution to get changes for.
param: geoids(character vector - default NULL) -> geoids to get changes for.
if none provided, returns all of US
return: dataframe of relationship data with "type_change" column.}
\usage{
get_2010_2020_bound_changes(res = "tract", geoids = NULL)
}
\description{
determines if/how census tract boundaries changed from 2010 to 2020. Adds
an additional "type_change" column to 2010-2020 tract relationship file
indicating whether a tract did not change ("same"), a tract was split but
bounds did not change ("split"), or a tracts bounds moved ("moved").
param: res(character - default "tract") -> resolution to get changes for.
param: geoids(character vector - default NULL) -> geoids to get changes for.
if none provided, returns all of US
return: dataframe of relationship data with "type_change" column.
}
